NEWTAG-526 Deprecate non-freemarker output controls
