+++ code/src/java/plugin/exporttokens/deprecated/EndManualWhitespaceToken.java
@@ -1,6 +1,6 @@
 /*
- * PaperInfoToken.java
- * Copyright 2003 (C) Devon Jones <soulcatcher@evilsoft.org>
+ * EndManualWhitespaceToken.java
+ * Copyright 2006 (C) James Dempsey <jdempsey@users.sourceforge.net>
  *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -16,23 +16,21 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
- * Created on December 15, 2003, 12:21 PM
+ * Created on 6/05/2006
  *
- * Current Ver: $Revision: 18611 $
- * Last Editor: $Author: thpr $
- * Last Edited: $Date: 2012-12-11 00:27:59 +0100 (Di, 11 Dez 2012) $
+ * $Id:  $
  *
  */
-package plugin.exporttokens;
+package plugin.exporttokens.deprecated;
 
-import pcgen.core.SettingsHandler;
 import pcgen.core.display.CharacterDisplay;
 import pcgen.io.ExportHandler;
 import pcgen.io.exporttoken.AbstractExportToken;
-import pcgen.util.Logging;
 
-//PAPERINFO
-public class InvalidTextToken extends AbstractExportToken
+/**
+ * Token that represents the end of author controlled whitespace
+ */
+public class EndManualWhitespaceToken extends AbstractExportToken
 {
 	/**
 	 * @see pcgen.io.exporttoken.Token#getTokenName()
@@ -40,7 +38,7 @@ public class InvalidTextToken extends AbstractExportToken
 	@Override
 	public String getTokenName()
 	{
-		return "INVALIDTEXT";
+		return "ENDMANUALWHITESPACE";
 	}
 
 	/**
@@ -50,20 +48,8 @@ public class InvalidTextToken extends AbstractExportToken
 	public String getToken(String tokenSource, CharacterDisplay display,
 		ExportHandler eh)
 	{
-		String sourceText = tokenSource.substring(12);
-		
-		if (sourceText.equals("TOHIT"))
-		{
-			return SettingsHandler.getInvalidToHitText(); 
-		}
-		else if (sourceText.equals("DAMAGE"))
-		{
-			return SettingsHandler.getInvalidDmgText();
-		}
-		else
-		{
-			Logging.errorPrint("Invalid INVALIDTEXT token:" + tokenSource);
-			return "";
-		}
+		eh.setManualWhitespace(false);
+		return "";
 	}
+
 }
