+++ code/src/java/plugin/exporttokens/deprecated/SpaceToken.java
@@ -1,6 +1,6 @@
 /*
- * BRToken.java
- * Copyright 2005 (C) James Dempsey <jdempsey@users.sourceforge.net>
+ * SpaceToken.java
+ * Copyright 2008 (C) PCGen
  *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -16,25 +16,25 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  *
- * Created on October 13, 2005, 8:20 PM
+ * Created on December 01, 2008, 12:21 PM
  *
- * Current Ver: $Revision$
- * Last Editor: $Author$
- * Last Edited: $Date$
- *
+ * Current Ver: $Revision: 1777 $
+ * Last Editor: $Author: jdempsey $
+ * Last Edited: $Date: 2006-12-17 04:36:01 +0000 (Sun, 17 Dec 2006) $
  */
-package plugin.exporttokens;
+package plugin.exporttokens.deprecated;
 
 import pcgen.core.display.CharacterDisplay;
 import pcgen.io.ExportHandler;
 import pcgen.io.exporttoken.AbstractExportToken;
 
 /**
- * BRToken outputs a line break to the output file.
- * 
- * Token syntax: BR
+ * SpaceToken - Used to provide a breaking space character (e.g. ASCII 32) 
+ * character) for use in a MANUASLWHITESPACE section as 
+ * the MANUALWHITESPACE token removes all whitespace and an HTML &nbsp; is 
+ * sometimes not desirable (as it doesn't naturally line break).
  */
-public class BRToken extends AbstractExportToken
+public class SpaceToken extends AbstractExportToken
 {
 	/**
 	 * @see pcgen.io.exporttoken.Token#getTokenName()
@@ -42,16 +42,16 @@ public class BRToken extends AbstractExportToken
 	@Override
 	public String getTokenName()
 	{
-		return "BR";
+		return "SPACE";
 	}
 
 	/**
-	 * @see pcgen.io.exporttoken.Token#getToken(java.lang.String, pcgen.core.PlayerCharacter, pcgen.io.ExportHandler)
+	 * @see pcgen.io.exporttoken.AbstractExportToken#getToken(java.lang.String, pcgen.core.display.CharacterDisplay, pcgen.io.ExportHandler)
 	 */
 	@Override
 	public String getToken(String tokenSource, CharacterDisplay display,
 		ExportHandler eh)
 	{
-		return "[BR]";
+		return " ";
 	}
 }
